+++ src/index.ts
@@ -1,3 +1,3 @@
-export { default as formatStats } from './formatStats';
+export { default as formatStats, Fallback } from './formatStats';
 export { default as formatValueRange } from './localize/formatValueRange';
 export { default as formatValue } from './localize/formatValue';